test: remove documentation tests; keep build/CI/.gitignore guards
Documentation does not need tests. Drops ~70 TestDocIntegrity_*,
TestDocSweep_*, TestDocs_*, and TestREADME_* tests that asserted on
markdown content under docs/, README.md, and CLAUDE.md. Splits the
remaining helpers and operational tests (CI workflow, Makefile,
.gitignore) into build_config_test.go.
